requirements.txt
requests
python-dotenv
PyYAML
google-genai
tqdm
beautifulsoup4
pydantic